feat(vocabulary): 添加获取词汇练习记录的功能及相关调整

+++ src/main/java/com/code/cetboot/dto/VocabularyPracticeDTO.java
@@ -28,5 +28,5 @@ public class VocabularyPracticeDTO {
             message = "词汇id不能为空",
             groups = {VocabularyPracticeValidation.Submit.class}
     )
-    private long vocabularyId;
+    private int vocabularyId;
 }
